Split in-memory-db-stubs into two source files
